tested error successfully

+++ src/data/explorer.py
@@ -1,6 +1,6 @@
 from .init import (conn, curs, IntegrityError)
 from ..model.explorer import Explorer
-from errors import Missing, Duplicate
+from ..errors import Missing, Duplicate
 
 curs.execute("""create table if not exists explorer(
              name text primary key,
@@ -8,7 +8,9 @@ curs.execute("""create table if not exists explorer(
              description text)""")
 
 def row_to_model(row:tuple)-> Explorer:
-    return Explorer(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    return Explorer(name = name,
+                    country = country, description = description)
 
 def model_to_dict(explorer: Explorer) -> dict:
     return explorer.model_dump() if explorer else None
@@ -18,7 +20,12 @@ def get_one(name: str) -> Explorer | None:
     params = {'name': name}
     curs.execute(qry, params)
     row = curs.fetchone()
-    return row_to_model(row) if row else None
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f'Explorer {name} not found')
+
+    # return row_to_model(row) if row else None
 
 def get_all() -> list[Explorer]:
     qry = 'select * from explorer'
@@ -26,11 +33,14 @@ def get_all() -> list[Explorer]:
     return [row_to_model(row) for row in curs.fetchall()]
 
 def create(explorer: Explorer) -> Explorer:
+    if not explorer: return None
     qry = """insert into explorer (name, country, description)
     values (:name, :country, :description) """
     params = model_to_dict(explorer)
-    curs.execute(qry,params)
-    conn.commit()
+    try:
+        curs.execute(qry,params)
+    except IntegrityError:
+        raise Duplicate(msg = f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
 def modify(name: str, explorer: Explorer) -> Explorer:
@@ -40,11 +50,17 @@ def modify(name: str, explorer: Explorer) -> Explorer:
              description =:description
              where name=:name_orig"""
     params = model_to_dict(explorer)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    params ['name_orig'] = explorer.name
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg = f"Explorer {name} not found")
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name: return False
     qry = 'delete from explorer where name =: name'
-    params = {'name': explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {'name':name}
+    curs.execute(qry, params)
+    if curs.rowcount !=1:
+        raise Missing(msg = f"Explorer {name} not found")
